% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_query.R
\name{execute_cypher}
\alias{execute_cypher}
\title{Title Execute a cypher (https://neo4j.com/developer/cypher/) query safely.}
\usage{
execute_cypher(cypher_query, verbose = FALSE)
}
\arguments{
\item{cypher_query}{A string of the cypher_query}

\item{verbose}{Boolean flag, wether or not verbose output should be provided.}
}
\value{
List datastructure containing the raw output of the response.
}
\description{
Title Execute a cypher (https://neo4j.com/developer/cypher/) query safely.
}
\examples{
raw_result = execute_cypher('MATCH (match_node:Match) RETURN match_node LIMIT 5')
}
