% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_query.R
\name{get_active_queries}
\alias{get_active_queries}
\title{Title Get the number of active queries currently running on the DB}
\usage{
get_active_queries()
}
\value{
Integer count of the active queries
}
\description{
Title Get the number of active queries currently running on the DB
}
