% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_query.R
\name{get_home_goals}
\alias{get_home_goals}
\title{Get the home goals for a particular match, selected by match id}
\usage{
get_home_goals(match_id)
}
\arguments{
\item{match_id}{}
}
\value{

}
\description{
Get the home goals for a particular match, selected by match id
}
