% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_query.R
\name{get_matches}
\alias{get_matches}
\title{Get matches from the DB}
\usage{
get_matches(expansions, max_n = 10)
}
\arguments{
\item{max_n}{The maximum number of games to return, for now limit set to NULL}
}
\value{
A data.frame of matches which matches query
}
\description{
Get matches from the DB
}
\examples{
get_matches(max_n=NULL) # get all matches
}
