% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_query.R
\name{get_matches}
\alias{get_matches}
\title{Get matches from the DB}
\usage{
get_matches(expansions, date = NULL, team = NULL, max_n = 10)
}
\arguments{
\item{expansions}{A list of wanted match expansions as defined in match_queries}

\item{team}{A string name, if passed will return matches for only this team}

\item{max_n}{The maximum number of games to return, for now limit set to NULL}
}
\value{
A data.frame of matches which matches query
}
\description{
Get matches from the DB
}
\examples{
basic_expansions = c('home_team', 'away_team', 'home_goals', 'away_goals')
get_matches(basic_expansions, max_n=NULL) # get all matches
get_matches(basic_expansions, team='Arsenal') # get Arsenal matches (max 10)
}
